requirements.txt
prometheus-client~=0.12.0
requests~=2.27.1
PyYAML~=6.0
pyVmomi~=7.0.3
pyvim~=3.0.2
